requirements.txt
argparse==1.4.0
PyYAML==5.3
requests==2.23.0